fix: make sure deleteAllVersions deletes just the entity when given a entity with no versions

+++ spring-versions-jpa/src/main/java/org/springframework/versions/impl/LockingAndVersioningRepositoryImpl.java
@@ -416,6 +416,12 @@ public class LockingAndVersioningRepositoryImpl<T, ID extends Serializable> impl
             throw new LockOwnerException("not lock owner");
         }
 
+        // issue #2039: if no version series has been established yet then delete the entity
+        if (isAncestralRoot(entity)) {
+            this.delete(entity);
+            return;
+        }
+
         String sql = "delete from ${entityClass} t where t.${ancestorRootId} = " + getAncestralRootId(entity);
 
         StringSubstitutor sub = new StringSubstitutor(getAttributeMap(entity.getClass()));
